+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 * Perform all compatibility checks before
   attempting to restore the machine state.
 * Added Clang support.
-* Added Arm Linux 32/64 bit GCC support.
+* Added Arm Linux (aarch64, armv7hf) GCC support.
 * Added Conan 2.0 support.
 * Replace Utils/base64.hpp with Conan package base64.
 * Replace Utils/md5.c with Conan package hash-library.